feat: add more functions to JNI interface

+++ mpv/src/jvmCommonMain/kotlin/dev/zt64/mpvkt/render/MpvRenderParam.jvmCommon.kt
@@ -1,6 +1,6 @@
 package dev.zt64.mpvkt.render
 
-public actual typealias MpvRenderParam = Long
+public actual typealias MpvRenderParam = Int
 
 public actual fun MpvRenderParam(name: String, value: String): MpvRenderParam {
     TODO()
